car-segmentation-ms/requirements.txt
# PyTorch — install the appropriate variant BEFORE this file (see README).
# GPU (CUDA 12.6): pip install torch==2.8.0 torchvision==0.23.0 torchaudio==2.8.0 --index-url https://download.pytorch.org/whl/cu126
# CPU only:        pip install torch==2.8.0 torchvision==0.23.0 torchaudio==2.8.0 --index-url https://download.pytorch.org/whl/cpu
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# ML / vision
numpy==1.26.4
opencv-python-headless==4.11.0.86
pillow==12.2.0
ultralytics==8.4.41
git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf

# API / server
fastapi==0.136.1
uvicorn[standard]==0.46.0
python-multipart==0.0.26
python-dotenv==1.2.2
openai==2.32.0
requests==2.33.1
charset-normalizer==3.4.7

# GCS — model weight download at startup
google-cloud-storage==3.10.1
